Change type of size in getStream from unsigned long to size_t, and add memcpy to postgresql stream

+++ src/iologindata.cpp
@@ -299,7 +299,7 @@ bool IOLoginData::loadPlayer(Player* player, DBResult_ptr result)
 	player->capacity = result->getNumber<uint32_t>("cap") * 100;
 	player->blessings = result->getNumber<uint16_t>("blessings");
 
-	uint64_t conditionsSize;
+	size_t conditionsSize;
 	const char* conditions = result->getStream("conditions", conditionsSize);
 	PropStream propStream;
 	propStream.init(conditions, conditionsSize);
@@ -946,7 +946,7 @@ void IOLoginData::loadItems(ItemMap& itemMap, DBResult_ptr result)
 		uint16_t type = result->getNumber<uint16_t>("itemtype");
 		uint16_t count = result->getNumber<uint16_t>("count");
 
-		uint64_t attrSize;
+		size_t attrSize;
 		const char* attr = result->getStream("attributes", attrSize);
 
 		PropStream propStream;
